+++ WordPress/src/main/java/org/wordpress/android/ui/main/MeFragment.kt
@@ -14,6 +14,7 @@ import android.text.TextUtils
 import android.view.View
 import android.view.View.OnClickListener
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.runtime.collectAsState
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
@@ -63,6 +64,7 @@ import org.wordpress.android.ui.deeplinks.DeepLinkOpenWebLinksWithJetpackHelper
 import org.wordpress.android.ui.jetpackoverlay.JetpackFeatureRemovalOverlayUtil
 import org.wordpress.android.ui.main.MeViewModel.RecommendAppUiState
 import org.wordpress.android.ui.main.WPMainActivity.OnScrollToTopListener
+import org.wordpress.android.ui.main.emailverificationbanner.EmailVerificationBanner
 import org.wordpress.android.ui.main.emailverificationbanner.EmailVerificationViewModel
 import org.wordpress.android.ui.main.utils.MeGravatarLoader
 import org.wordpress.android.ui.mysite.jetpackbadge.JetpackPoweredBottomSheetFragment
@@ -295,7 +297,16 @@ class MeFragment : Fragment(R.layout.me_fragment), OnScrollToTopListener {
 
         refreshWPCOMLoggedInOnlyButtonsVisibility()
 
-        meEmailVerificationBanner.setViewModel(emailVerificationViewModel)
+        meEmailVerificationBanner.setContent {
+            EmailVerificationBanner(
+                verificationState = emailVerificationViewModel.verificationState.collectAsState(),
+                emailAddress = emailVerificationViewModel.emailAddress.collectAsState(),
+                errorMessage = emailVerificationViewModel.errorMessage.collectAsState(),
+                onSendLinkClick = {
+                    emailVerificationViewModel.onSendVerificationLinkClick()
+                }
+            )
+        }
     }
 
     private fun MeFragmentBinding.refreshWPCOMLoggedInOnlyButtonsVisibility() {
